projects/agent/02-rag-agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
faiss-cpu
langchain==0.1.9
langchain-community==0.0.25
PyPDF2==3.0.1
python-multipart==0.0.6
aiosqlite==0.19.0
tqdm==4.66.1